requirements.txt
# Discord bot
py-cord==2.6.1            # slash commands etc. (this is the maintained fork)
discord-webhook>=1.3      # only if you're still posting via webhook URLs

# FreshRSS + helpers
freshrss-api==2.0.2
requests>=2.32

# XML parsing
lxml>=5.2

# Utilities
loguru>=0.7
python-dotenv>=1.0
